Music: 10 Block Stinger
This serves as a transition between the regular gameplay loop and the more intense gameplay loop. When a player has 10 or more blocks, this stinger should play once. Immediately after, the intense gameplay loop should kick in and loop (unless someone drops below 10 blocks, in which case we revert back to the regular intensity gameplay loop).